setup.py: Move "update_fa5" build command to a separate file called setupbase.py
